www: fix react-snap building
Need to investigate a bit more, but essentially when the site is crawled
if there is a request to the API and it fails, it was previously
throwing, but now we just log the error. This will be refined anyway so
that the client can see when things fail to load.