Add DTA logo for console emails
Note: this image will be used in HTML emails from the cloud.gov.au
console so this image should never be deleted. If it needs to be deleted
then a redirect needs to be put in place in the cloud.gov.au server in
order to keep old console emails working.